add modifyFriendDisplayName

+++ tux.cpp
@@ -3,6 +3,7 @@
 #include "userinformation.h"
 #include "utils.h"
 #include <QInputDialog>
+#include <QMessageBox>
 
 Tux::Tux(UserMessage userMessage, QWidget *parent) :
     QWidget(parent),
@@ -12,12 +13,14 @@ Tux::Tux(UserMessage userMessage, QWidget *parent) :
     ui->setupUi(this);
     ui->UserNameLabel->setText(userMsg.name);
 
-    api.getUserList(Utils::getInstance()->getTcpSocket(), friendList);
-    showFriendList();
-
+    timer.setInterval(30000);
+    refreshFriendList();
     setupMenu();
     connect(ui->listWidgetFriend, SIGNAL(customContextMenuRequested(QPoint)),
             this, SLOT(popupMenu(QPoint)));
+    connect(&timer, SIGNAL(timeout()), this, SLOT(refreshFriendList()));
+
+    timer.start();
 }
 
 Tux::~Tux()
@@ -25,8 +28,11 @@ Tux::~Tux()
     delete ui;
 }
 
-void Tux::showFriendList()
+void Tux::refreshFriendList()
 {
+    friendList.clear();
+    api.getUserList(Utils::getInstance()->getTcpSocket(), friendList);
+    ui->listWidgetFriend->clear();
     for(int i = 0; i < friendList.size(); ++i)
     {
         QString name;
@@ -42,11 +48,15 @@ void Tux::showFriendList()
 void Tux::setupMenu()
 {
     menu = new QMenu(this);
+    actionRefresh = menu->addAction("刷新好友列表");
     actionShowFriendProfile = menu->addAction("显示好友资料");
     actionModifyFriendDisplayName = menu->addAction("修改好友备注");
     actionDeleteFriend = menu->addAction("删除好友");
 
+    connect(actionRefresh, SIGNAL(triggered()), this, SLOT(refreshFriendList()));
     connect(actionDeleteFriend, SIGNAL(triggered()), this, SLOT(actionDeleteFriendTriggered()));
+    connect(actionModifyFriendDisplayName, SIGNAL(triggered()),
+            this, SLOT(actionModDisplayNameTriggered()));
 }
 
 void Tux::popupMenu(QPoint point)
@@ -61,10 +71,43 @@ void Tux::actionDeleteFriendTriggered()
     index = ui->listWidgetFriend->currentRow();
     if(index < 0)
         return;
+    QString name = friendList.at(index).nickName;
+    if(!friendList.at(index).displayName.trimmed().isEmpty())
+        name = friendList.at(index).displayName;
+    if(QMessageBox::question(this, "删除好友", QString("你确定要删除好友 %1？").arg(name),
+                             QMessageBox::Ok | QMessageBox::Cancel, QMessageBox::Cancel)
+            != QMessageBox::Ok)
+        return;
     if(api.deleteFriend(Utils::getInstance()->getTcpSocket(), friendList.at(index).account))
     {
         friendList.remove(index);
         delete ui->listWidgetFriend->takeItem(index);
+    }
+}
+
+void Tux::actionModDisplayNameTriggered()
+{
+    bool ok;
+    QString name;
+    int index;
+
+    index = ui->listWidgetFriend->currentRow();
+    if(index < 0)
+        return;
+
+    name = friendList.at(index).displayName;
+    name = QInputDialog::getText(this, "修改好友备注", "请输入备注名称：",
+                                 QLineEdit::Normal, name, &ok);
+    if(ok)
+    {
+        if(name.trimmed().isEmpty())
+        {
+            QMessageBox::information(this, "提示", "备注名称不能为空");
+            return;
+        }
+        if(api.modifyFriendDisplayName(Utils::getInstance()->getTcpSocket(),
+                                       friendList.at(index).account, name))
+            refreshFriendList();
     }
 }
